move more config related files into config specific folder
